ADD: Add manual sync
Add option to manually perform a synchronization between remote folder and local host. Available through the cmd-line command "sync", followed either by "from-spec <kernel_name>" or kernel start args (without connection file, as no kernel is started).

Pull functions `get_spec` and `check_spec` (from start.py) up to module level, and mark `check_spec` as private (`_check_spec`).
