+++ remote_kernel/start.py
@@ -1,4 +1,3 @@
-import argparse
 import json
 import logging
 import os
@@ -42,108 +41,6 @@ def generate_config():
     control_port=find_free_port(),
     transport='tcp'
   )
-
-
-def get_spec(argv=None):
-  global logger
-  from jupyter_core.paths import jupyter_path
-
-  spec_parser = argparse.ArgumentParser()
-  spec_parser.add_argument('kernel_name', help='Name of jupyter registered kernel to start (folder name)')
-  spec_args = spec_parser.parse_args(argv)
-
-  kernel_name = spec_args.kernel_name
-
-  if os.path.isabs(kernel_name):
-    kernel_spec = kernel_name
-  else:
-    kernel_spec = None
-    for d in ['.'] + jupyter_path('kernels'):
-      logger.debug('Searching in directory %s', d)
-      if os.path.isfile(os.path.join(d, kernel_name, 'kernel.json')):
-        kernel_spec = os.path.join(d, kernel_name, 'kernel.json')
-        break
-
-    assert kernel_spec is not None, \
-        'Kernel specification file %s not found!' % kernel_name
-
-  logger.info('Loading kernel specification file %s', kernel_spec)
-
-  with open(kernel_spec, mode='r') as spec_fs:
-    spec = json.load(spec_fs)
-
-  # Ensure specification is the correct version
-  spec = check_spec(kernel_spec, spec)
-
-  args = spec['argv']
-
-  assert args[1:3] == ['-m', 'remote_kernel'], \
-      'Kernel spec %s is not a remote_kernel specification' % kernel_spec
-
-  # Remove the jupyter supplied connection_file specification
-  if '-f' in args and args[args.index('-f') + 1] == '{connection_file}':
-    idx = args.index('-f')
-    del args[idx: idx+2]
-  elif '-f={connection_file}' in args:
-    del args[args.index('-f={connection_file}')]
-
-  return parse_args(args[3:])
-
-
-def check_spec(kernel_spec, spec):
-  """
-  Check if the deprecated "command" argument is present,
-  represents the first version of remote_kernel definition
-  If so, update to new-style kernel definition and re-save kernel spec
-  :param kernel_spec: File name where the kernel specification is stored
-  :param spec: JSON kernel specification to check and update if necessary
-  :return: Kernel specification, updated to new version if necessary.
-  """
-  global logger
-
-  args = spec['argv']
-  cmd = None
-  pre_args = None
-  post_args = None
-  if '--command' in args:
-    cmd_idx = args.index('--command')
-    cmd = args[cmd_idx + 1]
-    pre_args = args[:cmd_idx]
-    post_args = args[cmd_idx + 2:]
-  elif '-c' in args:
-    cmd_idx = args.index('-c')
-    cmd = args[cmd_idx + 1]
-    pre_args = args[:cmd_idx]
-    post_args = args[cmd_idx + 2:]
-  else:
-    for a_idx, a in enumerate(args):
-      if a.startswith('-c='):
-        cmd = a[3:]
-        pre_args = args[:a_idx]
-        post_args = args[a_idx + 1:]
-        break
-      elif a.startswith('--command='):
-        cmd = a[10:]
-        pre_args = args[:a_idx]
-        post_args = args[a_idx + 1:]
-        break
-
-  if cmd is not None:
-    cmds = cmd.split(' && ')
-    kernel_args = []
-    if len(cmds) > 1:
-      kernel_args += ['-pc', ' && '.join(cmds[:-1])]
-    if cmds[-1] != 'python -m ipykernel':
-      kernel_args += ['-k', cmds[-1]]
-
-    logger.warning('Deprecated version of kernel specification detected! Updating to new format...')
-    logger.debug('Normalizing command "%s" to %s', cmd, kernel_args)
-
-    spec['argv'] = pre_args + kernel_args + post_args
-
-    with open(kernel_spec, mode='w') as spec_fs:
-      json.dump(spec, spec_fs, indent=2)
-  return spec
 
 
 def parse_args(argv=None):
